Backend/requirements.txt
# Core Django and REST Framework
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
# Database
psycopg2-binary==2.9.11
dj-database-url==3.0.1
# Image handling
Pillow==12.0.0
qrcode==8.0
# Production server
gunicorn==23.0.0
# Environment variables
python-dotenv==1.1.1
# Testing
requests==2.32.3
# CORS support
django-cors-headers==4.3.1
